special enemy added/change name added

+++ TypingGame/Game.gd
@@ -1,11 +1,11 @@
 extends Node
 
-var Enemy = preload("res://Enemy.tscn")
-var Life = preload("res://Life.tscn")
+var Enemy = preload("res://Spawnable/Enemy.tscn")
+var SpecialEnemy = preload("res://Spawnable/SpecialEnemy.tscn")
+var Life = preload("res://Spawnable/Life.tscn")
 
 onready var enemy_container = $EnemyContainer
 onready var spawn_container = $SpawnContainer
-onready var spawn_timer = $SpawnTimer
 onready var difficulty_value = $MatchInformation/VBoxContainer/DifficultyRow/DiffucultyValue
 onready var player_Name_Label = $MatchInformation/VBoxContainer/PlayerScoreRow/PlayerNameLabel
 onready var player_score_value = $MatchInformation/VBoxContainer/PlayerScoreRow/PlayerScoreValue
@@ -18,7 +18,7 @@ onready var time_value = $MatchInformation/VBoxContainer/TimerRow/TimeValue
 onready var rematch_button = $gameOverScreen/PanelContainer/VBoxContainer/CenterContainer2/HBoxContainer/Rematch
 onready var life_container = $MatchInformation/VBoxContainer/LifeRow/LifeContainer
 onready var life_pos = $MatchInformation/VBoxContainer/LifeRow/lifePos
-
+onready var concede_button = $MatchInformation/VBoxContainer/ConcedeRow/concedeButton
 
 
 var current_letter_index=-1
@@ -64,7 +64,11 @@ func _unhandled_input(event:InputEvent)->void:
 		#ignore shift key press/hold
 		if typed_event.is_action_pressed("shift"):
 			return
-			
+		if typed_event.is_action_pressed("ui_select") and active_enemy != null:
+			active_enemy.reset_word_color()
+			active_enemy = null
+			return
+		
 		var key_typed = PoolByteArray([typed_event.unicode]).get_string_from_utf8()
 		if active_enemy == null:
 			find_new_active_enemy(key_typed)
@@ -78,7 +82,10 @@ func _unhandled_input(event:InputEvent)->void:
 				
 				if current_letter_index == prompt.length():
 					print("done")
-					word_killed()#add kill score
+					if active_enemy.is_special:
+						special_killed()
+					else:
+						word_killed()#add kill score
 					current_letter_index = -1
 					active_enemy.queue_free()
 					active_enemy = null
@@ -90,7 +97,9 @@ func word_killed():
 	Lobby.player_gain_score()
 	player_score_value.text = str(Lobby.player_score)
 
-
+func special_killed():
+	Lobby.player_gain_special_score()
+	player_score_value.text = str(Lobby.player_score)
 	
 func deduct_life():
 	Lobby.deduct_life_hit_bottom()
@@ -100,25 +109,12 @@ func deduct_life():
 		return
 	
 		
-	
-	
-
-
-#will be implement on server
-func _on_DifficultyTimer_timeout():
-#	difficulty += 1
-#	Global.emit_signal("difficulty_increased",difficulty)
-#	#print("difficulty increased to %d"%difficulty)
-#	var new_wait_time = spawn_timer.wait_time - 0.2
-#	spawn_timer.wait_time = clamp(new_wait_time,1,spawn_timer.wait_time)
-#	#timer won't be faster than 1 second
-#	difficulty_value.text = str(difficulty)
-	pass
-
 
 func _on_DeathArea_body_entered(body):
+	if not body.is_special:
+		deduct_life()
 	body.queue_free()
-	deduct_life()
+	
 	
 	
 func _on_concedeButton_pressed(): #timer in server should also stop
@@ -132,8 +128,12 @@ func handle_game_end():
 	game_over_screen.show()
 	remove_all_enemy()
 	
-func handle_spawn_enemy(word,spawn_index):
-	var enemy_instance = Enemy.instance()
+func handle_spawn_enemy(word,spawn_index,is_special):
+	var enemy_instance
+	if is_special:
+		enemy_instance = SpecialEnemy.instance()
+	else:
+		enemy_instance = Enemy.instance()
 	var spawns = spawn_container.get_children()
 	enemy_instance.set_difficulty(Lobby.difficulty)
 	enemy_instance.position = spawns[spawn_index].global_position
@@ -159,7 +159,7 @@ func _on_opponentLeftOkButton_pressed():
 	
 	
 func remove_all_enemy():
-	spawn_timer.stop()
+	concede_button.hide()
 	for enemy in enemy_container.get_children():
 		enemy.queue_free()
